+++ CMU_09_8220169_8220307_8220337/app/src/main/java/ipp/estg/cmu_09_8220169_8220307_8220337/data/room/models/leaderBoardEntries/LeaderboardEntryCalories.kt
@@ -2,5 +2,5 @@ package ipp.estg.cmu_09_8220169_8220307_8220337.data.room.models.leaderBoardEntr
 
 data class LeaderboardEntryCalories(
     val name: String,
-    val calories: Double
+    val calories: Int
 )
